+++ autoprecompiles/src/constraint_optimizer.rs
